+++ lib/column_api/objects/account_number.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module ColumnApi
+  class AccountNumber < Object
+  end
+end